Created Game Driver
Created a .NET project with some classes for the player objects, game instances, and the driver.

Driver: Starts a game instance
Game: Creates players, gets input for the number of ships, and handles the turns.
Player: Creates ships and places them on the board.
Shots are checked against the opposing player object's board array, and on hit, the opponents life is reduced.