Reworked Index.razor logic
Changed the game state indicator variables from booleans to an enum.

Coordinates are now stored in a tuple of type (int, int) instead of in two different integers.

Page now creates two boards, one for the player and their ships, another for that represents the enemy board and will hold the previous hits/misses.