Added ship plcement
Added the ability to place ships as well as view a preview of the ship being placed.

bool [,] previewBoard : A board that stores the true or false values based on whether or not the squares are currently showing a preview of the ship to be placed.

GetCellClass(int x, int y) : Changes the grid square based on certain criteria.

ship : A ship is on the current square
preview: A ship placement is being previewed on the current square
square: The default state (ie blank)

ShowShipPreview(int startX, int startY, bool isHorizontal): Sets the value of previewBoard to true based on what button is being hovered over, how long the ship is, and if it is horizontal/vertical.

HideShipPreview(int startX, int startY, bool isHorizontal): Hides the ship preview on mouseout.

PlaceShip(int startX, int startY, bool isHorizontal): Places a ship on mouse click based on start position, length of ship (Derived from shipCount) and whether or not the ship is horizontal. This function also increments ship count when it's called, and when ship count exceeds the shipLimit variable, changes the GameState to Playing.

enum GameState {Initial, Setup, Playing, Waiting, GameOver} : This enum holds the various game states.